+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+#7.0.11 (2019-05-22)
+
+- fix type of value returned for currency-edit-in-place : now return a number value, not a string of number
+- fix when removing all characters and submitting
+
 #7.0.10 (2019-05-21)
 
 - Add Edit-in-place with currency / fixed numbers